cmd/golink: add stat loading and saving to DB
add new DB.LoadStats and DB.SaveStats methods.  DB.SaveSaveStats takes
an incremental count of clicks since the last call.  This is a noop for
FileDB (the full count is still saved in the Link struct on disk), but
will allow for the SqliteDB to store stats in a separate table with
timestamps to display only recent clicks.

Change-Id: I18c7db871854923b873ae100c844940242a32bb3