cmd/golink: autofocus destination input if short exists
When a user visits a non-existing go link, it's reasonable to assume
that they may then want to create that link, so autofocus on the
destination input.

We could also autofocus the short input on initial page load without a
provided short name, but I think that's less of a safe assumption.

Change-Id: Id8e15e2f02f81d319eb782576e5e19f73c0ccee4