add /.export-stats endpoint to dump stats table
This is really bare bones, but at least lets us get the stats data out.
This also made it clear to me how brittle some of our tests are because
of how much global state we have in the top-level golink package. We
should probably move all of that into a server type or something similar
before too long.

Updates #168

Change-Id: Ia101a4a3005adb9118051b3416f5a64a4a45987d
Signed-off-by: Will Norris <will@tailscale.com>