golink: replace xsrftoken with Sec-Fetch-Site validation
Deprecate the use of xsrftokens entirely and replace them with requiring
clients to send [Sec-Fetch-Site](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Sec-Fetch-Site) fetch metadata
request headers that clarify the relationship between the initating and
requested origin.

Add tests for Sec-Golink behavior which continues to be permitted as a
substitute for the new CSRF protection as was for the previous one.

Signed-off-by: Patrick O'Doherty <patrick@tailscale.com>